Add initial Rust application setup
The commit entails setting up a new Rust application and building its initial setup. A '.gitignore' file was created to ignore the "/target" directory. This directory is usually generated while building the application and should not be included in the version control. The 'Cargo.lock' file was added that lists the exact versions of dependencies that the application needs to build correctly. The 'Cargo.toml' file, which specifies the necessary dependencies with their respective versions for the application, was also added. Furthermore, the 'main.rs' source file was included. This file contains the main logic of this application: it detects an existing process (AltServer.exe), terminates it and then restarts it.